backend/pyproject.toml
[project]
name = "backend"
version = "3.0.0"
description = "FastAPI Backend Gateway (Microservices)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"